Move notices to notice package following Notice class refactor
- adapt import statements
- finalize fields and variables
