+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/EntityMustBeUniqueNotice.java
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.notice.error;
+package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
 
-import org.mobilitydata.gtfsvalidator.usecase.notice.base.ErrorNotice;
-import org.mobilitydata.gtfsvalidator.usecase.port.ValidationResultRepository;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
 
 import java.io.IOException;
 
 public class EntityMustBeUniqueNotice extends ErrorNotice {
     private final String fieldName;
 
-    public EntityMustBeUniqueNotice(String filename, String fieldName, String entityId) {
+    public EntityMustBeUniqueNotice(final String filename, final String fieldName, final String entityId) {
         super(filename, E_020,
                 "Duplicate entity",
                 "Entity must be unique in file: " + filename + "found other entity with same value for " +
@@ -33,8 +33,7 @@ public class EntityMustBeUniqueNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
